Use a specific version of runtime, instead of using Gradle's variant resolution.
Previously we added the KotlinVersion attribute to all configurations in the module where SKIE was applied to.
That unfortunately breaks transitive dependencies, when other modules would depend on it, as those wouldn't have the attribute.
This change no longer relies on the attribute and instead targets an artifact for the specific version directly.